ExtraçaodeDados/requirements.txt
streamlit
pandas
plotly
matplotlib
openpyxl
pypdf
python-docx